pyproject.toml
[tool.poetry]
name = "phrase-database-cli"
version = "0.1.0"
description = ""
authors = ["Kamal Sacranie <kamalsjunkmail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.36"
urwid = "^2.1.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pip-autoremove = "^0.10.0"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
